fastapi-nicegui-folium/requirements.txt
affine==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
bidict==0.23.1
blinker==1.7.0
branca==0.7.1
cachelib==0.9.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
color-operations==0.1.3
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
decorator==5.1.1
docutils==0.19
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.109.2
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Cors==4.0.0
flask-restx==1.3.0
folium==0.16.0
fonttools==4.51.0
frozenlist==1.4.1
gcloud==0.18.3
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-core==2.4.1
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
idna==3.6
ifaddr==0.2.0
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipyleaflet==0.18.2
ipython==8.18.1
ipywidgets==8.1.2
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
localtileserver==0.10.2
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
morecantile==5.3.0
multidict==6.0.5
nicegui==1.4.20
numexpr==2.10.0
numpy==1.26.4
oauth2client==4.1.3
orjson==3.10.0
packaging==24.0
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
pscript==0.7.7
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.17.2
pyparsing==3.1.2
pyproj==3.6.1
pystac==1.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.0
python-multipart==0.0.9
python-socketio==5.11.2
pytz==2024.1
PyYAML==6.0.1
rasterio==1.3.9
referencing==0.34.0
requests==2.31.0
rio-cogeo==5.3.0
rio-tiler==6.4.5
rpds-py==0.18.0
rsa==4.9
scooby==0.9.2
server-thread==0.2.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
snuggs==1.4.7
stack-data==0.6.3
starlette==0.36.3
traitlets==5.14.2
traittypes==0.2.1
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
vbuild==0.8.2
watchfiles==0.21.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.2
widgetsnbextension==4.0.10
wsproto==1.2.0
xyzservices==2024.4.0
yarl==1.9.4
zipp==3.18.1
